✨ Now Introduced a Custom Hook
