+++ src/date-time-counter/CountDownTimer.js
@@ -1,29 +1,10 @@
-import { useEffect, useState } from "react";
 import "./datetime.css";
 import DateTimeDisplay from "./DateTimeDisplay";
+import useCountDown from "./hooks/useCountDown";
 
 const CountDownTimer = ({ targetDate }) => {
-  const countDownDate = new Date(targetDate).getTime();
-
-  const [countDown, setCountDown] = useState(
-    countDownDate - new Date().getTime()
-  );
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCountDown(countDownDate - new Date().getTime());
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, [countDownDate]);
-
-  // calculate time left
-  const days = Math.floor(countDown / (1000 * 60 * 60 * 24));
-  const hours = Math.floor(
-    (countDown % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)
-  );
-  const minutes = Math.floor((countDown % (1000 * 60 * 60)) / (1000 * 60));
-  const seconds = Math.floor((countDown % (1000 * 60)) / 1000);
+  
+  const [days, hours, minutes, seconds] = useCountDown(targetDate);
 
   return (
     <div className="countdown-container">
@@ -36,7 +17,7 @@ const CountDownTimer = ({ targetDate }) => {
         <DateTimeDisplay
           value={days}
           type={"Days"}
-          isDanger={days <= 1}
+          isDanger={days <= 3}
         />
         <p>:</p>
         <DateTimeDisplay
@@ -54,7 +35,7 @@ const CountDownTimer = ({ targetDate }) => {
         <DateTimeDisplay
           value={seconds}
           type={"Seconds"}
-          isDanger={true}
+          isDanger={false}
         />
       </a>
     </div>
